+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 ### Improvements
 
 - **Chip**: Fix demo page where there was a closable chip in the basic examples
+- **Table**:
+  - Add an option `noBackground` on both `pa-table-header` directive and `pa-table-sortable-header` component
+  - Add an option `border` on the table
 
 # 2.65.2 (2024-03-19)